+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [1.3.1] - 08-02-2018
+### Fixes
+- Fix incorrect message for checking client age
+
+### Added
+- Changed `reload` to `rconfig`
+- Added ability to reconfigure logger using `--logger` with `rconfig`
+
 ## [1.3.0] - 06-02-2018
 ### Added
 - New clients CLI option `clean`